mv hint() to command
